+++ app/controllers/timeline_controller.rb
@@ -1,4 +1,5 @@
 class TimelineController < ApplicationController
   def index
+    @continue_draft = Memory.most_recent_incomplete
   end
 end